Refactor API controller structure and Swagger config
Moved user-related controllers into a User namespace and directory for better organization. Unified Swagger documentation to a single API doc, updated tags and groupings for Admin and User endpoints, and improved route consistency. Adjusted controller attributes and tags to match the new structure. Minor code formatting and logic fixes in book management commands and handlers.
