+++ src/Booklify.API/Controllers/Admin/AuthController.cs
@@ -18,7 +18,7 @@ namespace Booklify.API.Controllers.Admin;
 /// </summary>
 [ApiController]
 [Route("api/cms/auth")]
-[ApiExplorerSettings(GroupName = "admin")]
+[ApiExplorerSettings(GroupName = "v1")]
 [Configurations.Tags("Admin", "Admin_Auth")]
 [SwaggerTag("API xác thực dành cho quản trị viên")]
 public class AuthController : ControllerBase
