+++ src/Booklify.API/Controllers/Admin/BookController.cs
@@ -347,7 +347,7 @@ public class BookController : ControllerBase
         var file = Request.Form.Files.GetFile("file");
         
         if (file != null)
-        {
+    {
             request.File = file;
         }
