+++ src/Booklify.Application/Features/Book/Commands/ManageBookStatus/ManageBookStatusCommand.cs
@@ -3,5 +3,5 @@ using Booklify.Application.Common.Models;
 using MediatR;
 
 namespace Booklify.Application.Features.Book.Commands.ManageBookStatus;
-
+ 
 public record ManageBookStatusCommand(Guid BookId, ManageBookStatusRequest Request) : IRequest<Result<BookResponse>>;
